Updates to test file and envisaclient
Just to be clear, the test file is how I test my scripts without having
to connect to the envisalink module.  These changes aren't really that
important.

Now that the objects file is ready for production, I'm updating
envisaclient.js.  Added a lot of comments to help understand what I'm
doing as well as minor changes so that this script works with the
objects module properly