Create app.js
So this is my new and improved envisaclient.js which includes a webserver function from socket.io.

Right now I can send commands from the web gui to the envisaclient, but I can't send the responses back to the web gui, mainly because "socket.io" can't be used outside the webserver function. I need to find a different route.